feat: item-formロジックを別ファイル切り出し、onsubmitもjsファイルに移行(#3)
