+++ index.html
@@ -29,11 +29,7 @@
       <h1 class="header__page-title">Checkout</h1>
     </header>
     <main class="container">
-      <form
-        action="#"
-        class="form-layout"
-        onsubmit="alert('submit success!'); return false;"
-      >
+      <form id="checkout-form" action="#" class="form-layout">
         <div>
           <div class="item-form">
             <ul class="item-list">
